recipe-v1/recipe_project/requirements.txt
django==3.0.6
djangorestframework==3.11.0
psycopg2==2.8.5
Pillow==7.1.2